Remove map, simplify.
< layout, interface
