+++ src/main/java/website/frontrow/board/Enemy.java
@@ -36,7 +36,7 @@ public class Enemy
     public Enemy(Point position, Map<Direction, Sprite> sprites)
     {
         super(true, position, new Point(0, 0), sprites);
-        addToLog("[BUBBLE]\t[SPAWN]\tEnemy created with ID: " + super.getId() + ".");
+        addToLog("[BUBBLE]\t[SPAWN]\tEnemy created.");
     }
 
     @Override
