+++ src/main/java/website/frontrow/board/Player.java
@@ -54,7 +54,7 @@ public class Player
     public Player(Point position, Map<Direction, Sprite> sprites)
     {
         super(true, position, new Point(0, 0), sprites);
-        addToLog("[PLAYER]\t[SPAWN]\tPlayer created with ID: " + super.getId() + ".");
+        addToLog("[PLAYER]\t[SPAWN]\tPlayer created with ID.");
     }
 
     /**
