+++ src/main/java/website/frontrow/board/Unit.java
@@ -23,16 +23,6 @@ public abstract class Unit
     private boolean alive;
 
     /**
-     * A unique id for the unit.
-     */
-    private final long id;
-
-    /**
-     * Count the amount of instantiations.
-     */
-    private static long count = 0;
-
-    /**
      * Current location of an entity.
      */
     @SuppressWarnings("visibilitymodifier") // subclasses have to have access to this variable
@@ -47,8 +37,6 @@ public abstract class Unit
     {
         this.alive = alive;
         this.location = location;
-        // This currently causes a write to a static variable fro an instance method.
-        this.id = this.count++;
     }
 
     /**
@@ -188,14 +176,5 @@ public abstract class Unit
     public double getSpeedMultiplier()
     {
     	return 1.0;
-    }
-
-    /**
-     * Get the id of the instantiation.
-     * @return Return a long with the id.
-     */
-    public long getId()
-    {
-        return id;
     }
 }
